backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
requests==2.31.0
numpy==1.26.4
httpx==0.26.0
python-dotenv==1.0.1

SQLAlchemy==2.0.30
alembic==1.13.1

passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
